// src/lib/discord/types.ts
export interface DiscordEmbed {
  title?: string;
  description?: string;
  url?: string;
  color?: number;
  fields?: Array<{
    name: string;
    value: string;
    inline?: boolean;
  }>;
  timestamp?: string;
  footer?: {
    text: string;
  };
}

export interface DiscordMessage {
  embeds: DiscordEmbed[];
  username?: string;
  avatar_url?: string;
}